refactor: compositions와 pages 분리
