+++ src/main.ts
@@ -1,6 +1,6 @@
 import { handleMovieSearch } from "./dom/eventHandler/handleMovieSearch";
-import { renderMainUI } from "./dom/render/renderMainUI";
-import { renderSearchUI } from "./dom/render/renderSearchUI";
+import { renderHomePage } from "./pages/home.ts";
+import { renderSearchPage } from "./pages/search.ts";
 
 const logo = document.getElementById("logo");
 const searchInput = document.getElementById(
@@ -24,15 +24,15 @@ if (searchInput && searchButton) {
   });
 }
 
-const render = async () => {
+const main = async () => {
   const url = new URL(window.location.href);
   const params = url.searchParams;
   const keyword = params.get("keyword");
   if (keyword) {
-    await renderSearchUI();
+    await renderSearchPage();
   } else {
-    await renderMainUI();
+    await renderHomePage();
   }
 };
 
-await render();
+await main();
